# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VSTR Management System API",
        "description": "Visitor Management System API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Server"
        },
        {
            "url": "https://api.vstrmgmt.com",
            "description": "Production Server"
        },
        {
            "url": "https://staging-api.vstrmgmt.com",
            "description": "Staging Server"
        }
    ],
    "paths": {
        "/admin/kiosks": {
            "get": {
                "tags": [
                    "Kiosks"
                ],
                "summary": "List kiosks",
                "description": "List kiosks for admin",
                "operationId": "listKiosks",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Kiosks"
                ],
                "summary": "Create kiosk",
                "description": "Create a new kiosk",
                "operationId": "createKiosk",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/kiosks/{kiosk}": {
            "get": {
                "tags": [
                    "Kiosks"
                ],
                "summary": "Show kiosk",
                "description": "Retrieve a kiosk",
                "operationId": "showKiosk",
                "parameters": [
                    {
                        "name": "kiosk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Kiosks"
                ],
                "summary": "Update kiosk",
                "description": "Update a kiosk",
                "operationId": "updateKiosk",
                "parameters": [
                    {
                        "name": "kiosk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/kiosks/{kiosk}/regenerate": {
            "post": {
                "tags": [
                    "Kiosks"
                ],
                "summary": "Regenerate kiosk credentials",
                "description": "Regenerate kiosk secret/credentials",
                "operationId": "regenerateKiosk",
                "parameters": [
                    {
                        "name": "kiosk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/kiosks/{kiosk}/revoke-tokens": {
            "post": {
                "tags": [
                    "Kiosks"
                ],
                "summary": "Revoke kiosk tokens",
                "description": "Revoke kiosk tokens",
                "operationId": "revokeKioskTokens",
                "parameters": [
                    {
                        "name": "kiosk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get current user",
                "description": "Returns authenticated user",
                "operationId": "getCurrentUser",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout",
                "description": "Logout authenticated user",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/session-check": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Session check",
                "description": "Check authenticated session",
                "operationId": "sessionCheck",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/kiosk/location-check": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Location check",
                "description": "Check location lock for kiosk",
                "operationId": "locationCheck",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/kiosk/me": {
            "get": {
                "tags": [
                    "Kiosk"
                ],
                "summary": "Get current kiosk user",
                "description": "Returns authenticated kiosk user",
                "operationId": "getKioskCurrentUser",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/kiosk/logout": {
            "post": {
                "tags": [
                    "Kiosk"
                ],
                "summary": "Logout kiosk",
                "description": "Logout kiosk session",
                "operationId": "logoutKiosk",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}